Move double_slider to widgets
