+++ src/view/widgets/double_slider.py
